Move files in dedicated folder
