+++ _dev/packages/mktg-with-google-common/translations/en/ui.json
@@ -629,7 +629,7 @@
 
     "productFeedPage": {
       "breadcrumb": {
-        "productStatusDetails": "Product status details",
+        "productStatusDetails": "Disapproved products",
         "nonCompliantProducts": "Non-compliant products"
       },
       "alert": {
@@ -744,17 +744,10 @@
         "enhancedFreeListing":  "Enhanced Free Listings"
       },
       "approvalTable": {
-        "description": "These are the status of your products that are sent to Google Merchant Center. Some products might appear more than once because of their language variation.",
+        "description": "Below is a list of your products that haven't passed the approval process. Each entry includes reasons for disapproval to help you make quick fixes. Once updated, these products can be sync to Google Shopping.",
         "filterTitle": "Filter by availability",
-        "filterAllStatus": "All",
-        "filterOnlyApproved": "Approved",
-        "filterOnlyPending": "Pending",
-        "filterOnlyNotApproved": "Not approved",
         "tableHeaderID": "ID",
-        "tableHeaderAttributeID": "Attribute",
-        "tableHeaderDestination": "Destination",
-        "tableHeaderName": "Title",
-        "tableHeaderGoogleValidation": "Availablity for Google",
+        "tableHeaderName": "product title",
         "tableHeaderLanguage": "Language",
         "tableHeaderCountry": "Country",
         "tableHeaderIssue": "Reason",
@@ -764,7 +757,7 @@
         "editX": "Edit {0}",
         "noReasonFallback": "*Not provided, check [account errors on your Merchant Center]({0})[:target=\"_blank\"].*"
       },
-      "productIssues": {
+      "productIssuesModal": {
         "title": "More details about your product issues",
         "description": "Here are the reasons for suspension provided by Google. Take the necessary steps to meet the requirements and wait for the next scan.",
         "footer": "Visit your [product sheet]({0}) to make the necessary changes to your product to meet Google's requirements and get it approved."
